Remove cyclic dependencies
