+++ core/src/main/scala/org/bykn/bosatsu/Parser.scala
@@ -44,18 +44,6 @@ object Parser {
     val toEOLIndent: Indy[Unit] =
       lift(toEOL) *> parseIndent
 
-    /**
-     * A: B or
-     * A:
-     *   B
-     */
-    def block[A, B](first: Indy[A], next: Indy[B]): Indy[(A, OptIndent[B])] =
-      blockLike(first, next, P(maybeSpace ~ ":"))
-
-    def blockLike[A, B](first: Indy[A], next: Indy[B], sep: P[Unit]): Indy[(A, OptIndent[B])] =
-      (first <* lift(sep ~/ maybeSpace))
-        .product(OptIndent.indy(next))
-
     implicit class IndyMethods[A](val toKleisli: Indy[A]) extends AnyVal {
       def region: Indy[(Region, A)] =
         toKleisli.mapF(_.region)
